feat!: support multiple gateways
When you have many devices on your network, more than one of them
may claim to be an Internet Gateway Device.

The `upnpNat` factory function now returns an object with a
`findGateways` method that returns an async generator which yields
`Gateway` objects which now contain the `map`/`unmap` and
`externalIp` methods.

BREAKING CHANGE: use `findGateways` to find a gateway before mapping ports
