+++ package.json
@@ -138,6 +138,7 @@
     "clean": "aegir clean",
     "lint": "aegir lint",
     "dep-check": "aegir dep-check",
+    "doc-check": "aegir doc-check",
     "build": "aegir build --no-bundle",
     "docs": "aegir docs",
     "release": "aegir release",
@@ -148,10 +149,9 @@
     "@achingbrain/ssdp": "^4.1.0",
     "@libp2p/logger": "^5.0.1",
     "abort-error": "^1.0.0",
-    "any-signal": "^4.1.1",
     "default-gateway": "^7.2.2",
     "err-code": "^3.0.1",
-    "it-first": "^3.0.1",
+    "it-merge": "^3.0.5",
     "p-defer": "^4.0.0",
     "race-signal": "^1.1.0",
     "xml2js": "^0.6.0"
