+++ src/upnp/constants.ts
@@ -1,0 +1,5 @@
+export const UPNP_ST = 'urn:schemas-upnp-org:device:InternetGatewayDevice:1'
+export const UPNP2_ST = 'urn:schemas-upnp-org:device:InternetGatewayDevice:2'
+export const DEFAULT_PORT_MAPPING_TTL = 720_000
+export const DEFAULT_REFRESH_TIMEOUT = 10_000
+export const DEFAULT_REFRESH_OFFSET = 60_000
