+++ test/pmp.spec.ts
@@ -1,20 +1,25 @@
 import { expect } from 'aegir/chai'
 import { pmpNat } from '../src/index.js'
-import type { NatAPI } from '../src/index.js'
+import { randomPort } from './fixtures/random-port.js'
+import type { Gateway, NatAPI } from '../src/index.js'
 
-describe('pmp-nat-port-mapper', () => {
+describe.skip('pmp-nat-port-mapper', () => {
   let client: NatAPI
+  let gateways: Gateway[]
 
-  before(() => {
+  beforeEach(() => {
     if (process.env.CI != null) {
       return // CI environments don't have uPNP routers!
     }
 
+    gateways = []
     client = pmpNat()
   })
 
-  after(async () => {
-    await client?.close()
+  afterEach(async () => {
+    await Promise.all(
+      gateways.map(async gateway => gateway.stop())
+    )
   })
 
   it('should map a port', async () => {
@@ -22,17 +27,21 @@ describe('pmp-nat-port-mapper', () => {
       return // CI environments don't have uPNP routers!
     }
 
-    const port = 48932
-    const mapped = await client.map(port)
+    const port = randomPort()
 
-    expect(mapped).to.be.a('number')
+    for await (const gateway of client.findGateways()) {
+      const mapped = await gateway.map(port)
 
-    process.on('SIGINT', () => {
-      void client.unmap(port)
-        .finally(() => {
-          process.exit(0)
-        })
-    })
+      expect(mapped).to.be.a('number')
+
+      await new Promise<void>((resolve) => {
+        setTimeout(() => {
+          resolve()
+        }, 5000)
+      })
+
+      await gateway.unmap(port)
+    }
   })
 
   it('should discover an external ip address', async () => {
@@ -40,10 +49,12 @@ describe('pmp-nat-port-mapper', () => {
       return // CI environments don't have uPNP routers!
     }
 
-    const ip = await client.externalIp({
-      signal: AbortSignal.timeout(5000)
-    })
+    for await (const gateway of client.findGateways()) {
+      const ip = await gateway.externalIp({
+        signal: AbortSignal.timeout(5000)
+      })
 
-    expect(ip).to.be.ok()
+      expect(ip).to.be.ok()
+    }
   })
 })
